Revert "Merge branch 'develop' into Jp-rerun"
This reverts commit e0ceaa2b887564a4f1fbe97112846d4ca62a1aa2, reversing
changes made to 5e9e4e1fc280632719886300dfcec73ff1120c7d.

+++ mm/src/code/speed_meter.c
@@ -1,16 +1,8 @@
-#include "prevent_bss_reordering.h"
-#include "z64speed_meter.h"
-
-#include "gfx.h"
-#include "regs.h"
+#include "global.h"
 #include "sys_cfb.h"
-#include "system_malloc.h"
-#include "z64game.h"
-#include "z64malloc.h"
+#include "z64speed_meter.h"
 #include "z64view.h"
-
-#include "functions.h"
-#include "variables.h"
+#include "system_malloc.h"
 
 /**
  * How much time the RSP ran audio tasks for over the course of `gGraphUpdatePeriod`.
@@ -223,9 +215,9 @@ void SpeedMeter_DrawAllocEntries(SpeedMeter* meter, GraphicsContext* gfxCtx, Gam
     TwoHeadArena* tha;
     s32 y;
     TwoHeadGfxArena* thga;
-    size_t zeldaFreeMax;
-    size_t zeldaFree;
-    size_t zeldaAlloc;
+    u32 zeldaFreeMax;
+    u32 zeldaFree;
+    u32 zeldaAlloc;
     s32 sysFreeMax;
     s32 sysFree;
     s32 sysAlloc;
@@ -243,7 +235,7 @@ void SpeedMeter_DrawAllocEntries(SpeedMeter* meter, GraphicsContext* gfxCtx, Gam
     }
 
     if (R_ENABLE_ARENA_DBG > 1) {
-        SystemArena_GetSizes((size_t*)&sysFreeMax, (size_t*)&sysFree, (size_t*)&sysAlloc);
+        SystemArena_GetSizes((u32*)&sysFreeMax, (u32*)&sysFree, (u32*)&sysAlloc);
         SpeedMeter_InitAllocEntry(&entry, sysFree + sysAlloc - state->tha.size, sysAlloc - state->tha.size,
                                   GPACK_RGBA5551(0, 0, 255, 1), GPACK_RGBA5551(255, 128, 128, 1), ulx, lrx, y, y);
         SpeedMeter_DrawAllocEntry(&entry, gfxCtx);
