+++ mm/src/code/z_skin_matrix.c
@@ -1,9 +1,4 @@
-#include "z64skin_matrix.h"
-
-#include "gfx.h"
-#include "macros.h"
-#include "z64math.h"
-
+#include "global.h"
 #include "2s2h/Enhancements/FrameInterpolation/FrameInterpolation.h"
 
 MtxF sMtxFClear = { {
@@ -540,7 +535,7 @@ void SkinMatrix_SetRotateAroundVec(MtxF* mf, s16 a, f32 x, f32 y, f32 z) {
     f32 xy;
     f32 yz;
     f32 xz;
-    s32 pad;
+    f32 pad;
 
     sinA = Math_SinS(a);
     cosA = Math_CosS(a);
